+++ sandbox/scheduler-example/src/datasources/schedular.cache.datasource.ts
@@ -1,0 +1,32 @@
+import {inject, lifeCycleObserver, LifeCycleObserver} from '@loopback/core';
+import {juggler} from '@loopback/repository';
+
+const config = {
+  name: 'AuthCache',
+  connector: 'kv-redis',
+  url: '',
+  host: process.env.REDIS_HOST,
+  port: process.env.REDIS_PORT,
+  password: process.env.REDIS_PASSWORD,
+  db: process.env.REDIS_DB,
+};
+
+// Observe application's life cycle to disconnect the datasource when
+// application is stopped. This allows the application to be shut down
+// gracefully. The `stop()` method is inherited from `juggler.DataSource`.
+// Learn more at https://loopback.io/doc/en/lb4/Life-cycle.html
+@lifeCycleObserver('datasource')
+export class AuthCacheDataSource
+  extends juggler.DataSource
+  implements LifeCycleObserver
+{
+  static dataSourceName = 'AuthCache';
+  static readonly defaultConfig = config;
+
+  constructor(
+    @inject('datasources.config.AuthCache', {optional: true})
+    dsConfig: object = config,
+  ) {
+    super(dsConfig);
+  }
+}