+++ sandbox/scheduler-example/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## <small>0.1.4 (2021-08-02)</small>
+
+* docs(schedular-service and video-conferencing-service): update README.md (#277) ([1257784](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/1257784)), closes [#277](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/277)
+
+
+
+
+
 ## <small>0.1.3 (2021-07-24)</small>
 
 * fix(core): upgrade loopback4-soft-delete ([f653e0c](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/f653e0c)), closes [#0](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/0)