+++ services/scheduler-service/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## <small>1.2.2 (2021-08-02)</small>
+
+* fix(scheduler-service): update openapi for video and scheduler service ([836e291](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/836e291)), closes [#0](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/0)
+* docs(schedular-service and video-conferencing-service): update README.md (#277) ([1257784](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/1257784)), closes [#277](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/277)
+
+
+
+
+
 ## <small>1.2.1 (2021-07-24)</small>
 
 * fix(core): upgrade loopback4-soft-delete ([f653e0c](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/f653e0c)), closes [#0](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/0)